Module aws_sdk_codestarnotifications::error

Modules

access_denied_exception

See AccessDeniedException

concurrent_modification_exception

See ConcurrentModificationException

configuration_exception

See ConfigurationException

invalid_next_token_exception

See InvalidNextTokenException

limit_exceeded_exception

See LimitExceededException

resource_already_exists_exception

See ResourceAlreadyExistsException

resource_not_found_exception

See ResourceNotFoundException

validation_exception

See ValidationException

Structs

AccessDeniedException

AWS CodeStar Notifications can't create the notification rule because you do not have sufficient

ConcurrentModificationException

AWS CodeStar Notifications can't complete the request because the resource is being modified by

ConfigurationException

Some or all of the configuration is incomplete, missing, or not valid.

CreateNotificationRuleError

DeleteNotificationRuleError

DeleteTargetError

DescribeNotificationRuleError

InvalidNextTokenException

The value for the enumeration token used in the request to return the next batch of the results is not valid.

LimitExceededException

One of the AWS CodeStar Notifications limits has been exceeded. Limits apply to

ListEventTypesError

ListNotificationRulesError

ListTagsForResourceError

ListTargetsError

ResourceAlreadyExistsException

A resource with the same name or ID already exists. Notification rule names must be

ResourceNotFoundException

AWS CodeStar Notifications can't find a resource that matches the provided ARN.

SubscribeError

TagResourceError

UnsubscribeError

UntagResourceError

UpdateNotificationRuleError

ValidationException

One or more parameter values are not valid.

Enums

CreateNotificationRuleErrorKind

DeleteNotificationRuleErrorKind

DeleteTargetErrorKind

DescribeNotificationRuleErrorKind

ListEventTypesErrorKind

ListNotificationRulesErrorKind

ListTagsForResourceErrorKind

ListTargetsErrorKind

SubscribeErrorKind

TagResourceErrorKind

UnsubscribeErrorKind

UntagResourceErrorKind

UpdateNotificationRuleErrorKind