Module aws_sdk_codestarnotifications::model

Modules

event_type_summary

See EventTypeSummary

list_event_types_filter

See ListEventTypesFilter

list_notification_rules_filter

See ListNotificationRulesFilter

list_targets_filter

See ListTargetsFilter

notification_rule_summary

See NotificationRuleSummary

target

See Target

target_summary

See TargetSummary

Structs

EventTypeSummary

Returns information about an event that has triggered a notification rule.

ListEventTypesFilter

Information about a filter to apply to the list of returned event types. You can filter

ListNotificationRulesFilter

Information about a filter to apply to the list of returned notification rules. You can

ListTargetsFilter

Information about a filter to apply to the list of returned targets. You can filter by

NotificationRuleSummary

Information about a specified notification rule.

Target

Information about the SNS topics associated with a notification rule.

TargetSummary

Information about the targets specified for a notification rule.

Enums

DetailType

ListEventTypesFilterName

ListNotificationRulesFilterName

ListTargetsFilterName

NotificationRuleStatus

TargetStatus