Module aws_sdk_codestarnotifications::operation

Structs

CreateNotificationRule

Creates a notification rule for a resource. The rule specifies the events you want

DeleteNotificationRule

Deletes a notification rule for a resource.

DeleteTarget

Deletes a specified target for notifications.

DescribeNotificationRule

Returns information about a specified notification rule.

ListEventTypes

Returns information about the event types available for configuring notifications.

ListNotificationRules

Returns a list of the notification rules for an AWS account.

ListTagsForResource

Returns a list of the tags associated with a notification rule.

ListTargets

Returns a list of the notification rule targets for an AWS account.

Subscribe

Creates an association between a notification rule and an SNS topic so that the

TagResource

Associates a set of provided tags with a notification rule.

Unsubscribe

Removes an association between a notification rule and an Amazon SNS topic so that

UntagResource

Removes the association between one or more provided tags and a notification

UpdateNotificationRule

Updates a notification rule for a resource. You can change the events that trigger the